securecookie: add package for ChaCha20Poly1305-based secure cookies
This package allows encrypting and decrypting an arbitrary type into a
string that can be used in a HTTP cookie. The encrypted value is
protected with an AEAD (see [1]), which means that it is both
confidential (i.e. the end user cannot see the value in the cookie) and
authenticated (i.e. the end user cannot modify the cookie's value).

[1]: https://en.wikipedia.org/wiki/Authenticated_encryption

Updates #4