integration: add integration tests for Caddy
This adds an integration test if the 'caddy' binary is available in the
$PATH that launches Caddy with forward_auth configured and validates
that logging in and then accessing a protected resource succeeds.

Additionally, this adds a 'testproc' helper library that can be used to
run a process during a test, managing its lifecycle and log output and
ensuring that it gets cleaned up when the test exits.

Updates #7