cmd/homeauth: stop using securecookie for WebAuthn session data
As @zofrex helpfully pointed out, we don't actually need to use a cookie
here, since we can just encrypt/authenticate the WebAuthn session data
and have the client pass it back to us using regular HTTP requests. This
simplifies a few things, and in general the fewer cookies that we use,
the better.

Updates #4

Signed-off-by: Andrew Dunham <andrew@du.nham.ca>