cmd/homeauth: magic login links should use POST
Now, when a user clicks on a magic login link, we render a HTML page
that contains a form that immediately submits itself, which makes a POST
request to actually complete the login.

This prevents an overzealous email client or security scanner from
"clicking" on the URL, making a GET request, and then invalidating the
token before a user can use it.

Updates #32

Signed-off-by: Andrew Dunham <andrew@du.nham.ca>